+++ db-conf/connect.js
@@ -1,0 +1,10 @@
+const mysql = require('mysql');
+
+const connection = mysql.createConnection({
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME
+});
+
+module.exports = connection;
